custom_1.py & custom_2.py assemble
