+++ main/custom_1.py
@@ -1,6 +1,6 @@
 import sys
 from PySide2.QtWidgets import (
-    QApplication, QWidget, QLabel, QPushButton, QVBoxLayout, QHBoxLayout,
+    QApplication, QWidget, QLabel, QPushButton, QVBoxLayout, QHBoxLayout, QTextBrowser,
     QGroupBox, QLineEdit, QGraphicsDropShadowEffect, QGraphicsOpacityEffect
 )
 from PySide2.QtCore import QPropertyAnimation
@@ -27,45 +27,21 @@ BUTTON_STYLE = """
     }
 """
 
-# 1. 메모리 사이즈 선택 박스
-class MemorySizeGroup(QGroupBox):
-    def __init__(self, title):
-        super().__init__(title)
-        self.setMaximumHeight(90)
-        self.setStyleSheet(self._groupbox_style())
-
-        layout = QHBoxLayout(spacing=15)
-        layout.setContentsMargins(10, 10, 10, 10)
-
-        for size in ["2KByte", "8KByte", "16KByte"]:
-            btn = QPushButton(size)
-            btn.setFixedSize(100, 40)
-            btn.setStyleSheet(BUTTON_STYLE)
-            layout.addWidget(btn)
-
-        self.setLayout(layout)
-
-    def _groupbox_style(self):
-        return """
-            QGroupBox {
-                font-weight: bold;
-                font-size: 14px;
-                border: 1px solid #888;
-                border-radius: 10px;
-                padding: 8px;
-            }
-        """
-
-
-# 2. 숫자 입력 박스 (공통 구조)
+# -----------------------------
+# 공통 입력 박스 (리팩토링된 버전)
 class IntegerInputGroup(QGroupBox):
-    def __init__(self, title, example_text):
+    def __init__(self, title, example_text, on_apply=None, notice_text=None):
         super().__init__(title)
-        self.setMaximumHeight(90)
+        self.on_apply_callback = on_apply
         self.setStyleSheet(self._groupbox_style())
+        self.setMaximumHeight(130)
 
-        layout = QHBoxLayout(spacing=10)
-        layout.setContentsMargins(10, 10, 10, 10)
+        main_layout = QVBoxLayout()
+        main_layout.setContentsMargins(10, 10, 10, 10)
+        main_layout.setSpacing(6)
+
+        h_layout = QHBoxLayout()
+        h_layout.setSpacing(10)
 
         self.input = QLineEdit()
         self.input.setPlaceholderText(example_text)
@@ -76,17 +52,34 @@ class IntegerInputGroup(QGroupBox):
         self.apply_btn = QPushButton("Apply")
         self.apply_btn.setFixedSize(80, 35)
         self.apply_btn.setStyleSheet(BUTTON_STYLE)
-        # self.apply_btn.clicked.connect(lambda: print(f"{title} applied:", self.input.text()))
         self.apply_btn.clicked.connect(self.save_category_num)
 
-        layout.addWidget(self.input)
-        layout.addWidget(self.apply_btn)
-        self.setLayout(layout)
+        h_layout.addWidget(self.input)
+        h_layout.addWidget(self.apply_btn)
+        main_layout.addLayout(h_layout)
+
+        if notice_text:
+            notice = QLabel(notice_text)
+            notice.setWordWrap(True)
+            notice.setStyleSheet("font-size: 11px; color: #555;")
+            main_layout.addWidget(notice)
+
+        self.setLayout(main_layout)
 
     def save_category_num(self):
-        self.category_num = int(self.input.text())
-        print(f"applied:", self.input.text())
+        try:
+            self.category_num = int(self.input.text())
+        except ValueError:
+            self.category_num = 0
+        print(f"applied: {self.category_num}")
+        if self.on_apply_callback:
+            self.on_apply_callback()
 
+    def get_value(self):
+        try:
+            return int(self.input.text())
+        except ValueError:
+            return 0
 
     def _groupbox_style(self):
         return """
@@ -109,34 +102,73 @@ class IntegerInputGroup(QGroupBox):
             }
         """
 
-
-# 3. Training dataset + 안내 텍스트
+# -----------------------------
+# 입력 박스 서브 클래스
 class TrainingInputGroup(IntegerInputGroup):
-    def __init__(self):
-        super().__init__("3. Number of training dataset", "ex) 1000       Write only one unsigned integer number")
-        self.setMaximumHeight(130)
-
-        notice = QLabel(
-            "\n".join([
-                "※ Number of training dataset should be less or equal than number of category.",
+    def __init__(self, on_apply=None):
+        super().__init__(
+            "2. Number of training dataset",
+            "ex) 1000       Write only one unsigned integer number",
+            on_apply=on_apply,
+            notice_text=(
+                "※ Number of training dataset should be more or equal than number of category to train.\n"
                 "※ We recommend preparing at least 100 samples for each category in the training dataset."
-            ])
+            )
         )
-        notice.setWordWrap(True)
-        notice.setStyleSheet("font-size: 11px; color: #555;")
 
-        layout = QVBoxLayout()
-        layout.addLayout(self.layout())
-        layout.addWidget(notice)
-        self.setLayout(layout)
+class InputVectorGroup(IntegerInputGroup):
+    def __init__(self, on_apply=None):
+        super().__init__(
+            "3. Input vector length",
+            "Write only one unsigned integer number",
+            on_apply=on_apply,
+            notice_text="※ Number of training dataset should be more or equal than number of category to train."
+        )
+
+# -----------------------------
+# 메모리 계산 박스
+class MemorySizeWindow(QGroupBox):
+    def __init__(self):
+        super().__init__("4. Required memory size")
+        self.setStyleSheet("""
+            QGroupBox {
+                font-weight: bold;
+                font-size: 14px;
+                border: 1px solid #888;
+                border-radius: 10px;
+                padding: 8px;
+            }
+        """)
+
+        self.layout = QVBoxLayout()
+        self.output_box = QTextBrowser()
+        self.output_box.setStyleSheet("""
+            QTextBrowser {
+                border: none;
+                padding: 5px;
+                background-color: #f9f9f9;
+            }
+        """)
+        self.layout.addWidget(self.output_box)
+        self.setLayout(self.layout)
+
+    def update_display(self, input_vector_length: int, training_dataset: int):
+        if input_vector_length <= 0 or training_dataset <= 0:
+            self.output_box.setText("Please enter valid input vector length and number of training dataset.")
+            return
+        memory_size = input_vector_length * training_dataset
+        self.output_box.setText(f"""
+input vector length × number of training dataset ≤ memory size\n
+⇔ {input_vector_length}Kbyte × {training_dataset} ≤ memory size\n
+∴ available memory size ≥ {memory_size}Kbyte""")
 
 
-# 4. 메인 SubWindow
+# -----------------------------
+# 메인 윈도우
 class Custom_1_Window(QWidget):
     def __init__(self):
         super().__init__()
-        self._setup_ui()       
-        
+        self._setup_ui()
 
     def _setup_ui(self):
         self.setWindowFlags(Qt.FramelessWindowHint | Qt.Window)
@@ -156,16 +188,27 @@ class Custom_1_Window(QWidget):
 
         layout = QVBoxLayout(container)
         layout.setContentsMargins(20, 60, 20, 20)
-        layout.setSpacing(50)
+        layout.setSpacing(40)
 
-        layout.addWidget(MemorySizeGroup("1. Intellino memory size"))
-        self.category_input = IntegerInputGroup("2. Number of category to train", "ex) 10       Write only one unsigned integer number")
+        self.category_input = IntegerInputGroup("1. Number of category to train", "ex) 10", self.update_memory_display)
         layout.addWidget(self.category_input)
-        layout.addWidget(TrainingInputGroup())
+
+        self.train_data_input = TrainingInputGroup(on_apply=self.update_memory_display)
+        layout.addWidget(self.train_data_input)
+
+        self.input_vector_input = InputVectorGroup(on_apply=self.update_memory_display)
+        layout.addWidget(self.input_vector_input)
+
+        self.memory_display = MemorySizeWindow()
+        layout.addWidget(self.memory_display)
 
         layout.addStretch()
         layout.addLayout(self._create_next_button())
-        
+
+    def update_memory_display(self):
+        vec_len = self.input_vector_input.get_value()
+        train_num = self.train_data_input.get_value()
+        self.memory_display.update_display(vec_len, train_num)
 
     def _add_title_bar(self, parent):
         title_bar = QWidget(parent)
@@ -183,7 +226,7 @@ class Custom_1_Window(QWidget):
         close_btn.setIcon(QIcon("main/home.png"))
         close_btn.setIconSize(QSize(24, 24))
         close_btn.setFixedSize(34, 34)
-        close_btn.setStyleSheet(("""
+        close_btn.setStyleSheet("""
             QPushButton {
                 border: none;
                 background-color: transparent;
@@ -192,7 +235,7 @@ class Custom_1_Window(QWidget):
                 background-color: #dee2e6;
                 border-radius: 17px;
             }
-            """))
+        """)
         close_btn.clicked.connect(self.close)
 
         layout.addWidget(logo_label)
@@ -228,16 +271,14 @@ class Custom_1_Window(QWidget):
             self.offset = event.pos()
 
     def mouseMoveEvent(self, event: QMouseEvent):
-
         if hasattr(self, 'offset') and event.buttons() == Qt.LeftButton:
             self.move(self.pos() + event.pos() - self.offset)
 
     def nextFunction(self):
-        num = self.category_input.category_num
+        num = self.category_input.get_value()
         self.custom2 = InputVectorWindow(num_categories=num)
         self.custom2.show()
 
-        # 페이드아웃 기능
         effect = QGraphicsOpacityEffect(self)
         self.setGraphicsEffect(effect)
         self.anim = QPropertyAnimation(effect, b"opacity")
@@ -246,7 +287,6 @@ class Custom_1_Window(QWidget):
         self.anim.setEndValue(0)
         self.anim.finished.connect(self.close)
         self.anim.start()
-        
 
 
 if __name__ == "__main__":
